Extract RC4 code to separate file
