+++ src/libmdb/Makefile.am
@@ -1,5 +1,5 @@
 lib_LTLIBRARIES	=	libmdb.la
-libmdb_la_SOURCES=	catalog.c file.c table.c data.c dump.c backend.c money.c sargs.c index.c like.c write.c stats.c map.c props.c worktable.c options.c iconv.c version.c
+libmdb_la_SOURCES=	catalog.c file.c table.c data.c dump.c backend.c money.c sargs.c index.c like.c write.c stats.c map.c props.c worktable.c options.c iconv.c version.c rc4.c
 libmdb_la_LDFLAGS = -version-info $(VERSION_INFO)
 if FAKE_GLIB
 libmdb_la_SOURCES += fakeglib.c
